requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.0.10
langchain-core>=0.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
dotenv>=0.0.5
fastapi
uvicorn
pydantic